+++ api/src/@types/database/ILoader.d.ts
@@ -1,3 +1,4 @@
 interface ILoader {
+    beforeRun(): Promise<boolean>;
     run(): Promise<void>;
 }