+++ api/src/@types/database/entities/items/IItem.d.ts
@@ -1,12 +1,16 @@
 interface IItem {
     id: number;
     name: string;
+    type: 'floor' | 'wall';
     width: number;
-    height: number;
     length: number;
+    height: number;
+    can_stack: boolean;
     can_walk: boolean;
     can_sit: boolean;
+    can_lay: boolean;
     interaction_count: number;
+    is_rare: boolean;
 
     catalog_items: ICatalogItem[];
     rooms: IRoom[];