+++ api/src/database/loaders/navigator/NavigatorCategoriesLoader.ts
@@ -2,13 +2,18 @@ import LogsManager from "@Logs";
 import SHabbo from "@SHabbo";
 
 class NavigatorCategoriesLoader implements ILoader {
+    async beforeRun(): Promise<boolean> {
+        const navigator_categories = await SHabbo.getDatabase().getNavigatorCategory().find({ take: 1 });
+        return navigator_categories.length === 0;
+    }
+
     async run(): Promise<void> {
         await SHabbo.getDatabase().getNavigatorCategory().create({
             id: 1,
             name: 'Street',
         }).save();
 
-        LogsManager.success('Inserted 1 navigator category.');
+        LogsManager.success('Default navigator categories inserted.');
     }
 }