+++ api/src/database/loaders/users/UsersLoader.ts
@@ -2,13 +2,18 @@ import LogsManager from "@Logs";
 import SHabbo from "@SHabbo";
 
 class UsersLoader implements ILoader {
+    async beforeRun(): Promise<boolean> {
+        const users = await SHabbo.getDatabase().getUsers().find({ take: 1 });
+        return users.length === 0;
+    }
+
     async run(): Promise<void> {
         await SHabbo.getDatabase().getUsers().create({
             id: 1,
             username: 'SHabbo',
         }).save();
 
-        LogsManager.success('Inserted 1 user.');
+        LogsManager.success('Default users inserted.');
     }
 }